ibv CAS for local aggregator

+++ ptre/communication/rdma/rdma_agg_writer.h
@@ -22,7 +22,7 @@ class RdmaAggWriter {
                 const std::vector<struct ibv_mr*>& send_buf_mrs);
   /// Returns 0 on success, -1 on failure.
   int TransitState(const string& name);
-  int TransitStateV2(const string& name, const uint64_t from,
+  uint64_t TransitStateV2(const string& name, const uint64_t from,
       const uint64_t to);
   int WriteToAggBuf(const string& name);
   int WriteToAggBufV2(const string& name);
